split into separate files and polish
